renaming login file - component based
